move class member functions to OpUtils
